MAINT: Move module to be `np.dtypes` and add release note
