+++ numpy/core/__init__.py
@@ -143,11 +143,11 @@ def _DType_reconstruct(scalar_type):
 
 def _DType_reduce(DType):
     # As types/classes, most DTypes can simply be pickled by their name:
-    if not DType._legacy or DType.__module__ == "numpy.types":
+    if not DType._legacy or DType.__module__ == "numpy.dtypes":
         return DType.__name__
 
     # However, user defined legacy dtypes (like rational) do not end up in
-    # `numpy.types` as module and do not have a public class at all.
+    # `numpy.dtypes` as module and do not have a public class at all.
     # For these, we pickle them by reconstructing them from the scalar type:
     scalar_type = DType.type
     return _DType_reconstruct, (scalar_type,)
