+++ numpy/dtypes.pyi
