+++ numpy/tests/test_public_api.py
@@ -136,6 +136,7 @@ PUBLIC_MODULES = ['numpy.' + s for s in [
     "doc",
     "doc.constants",
     "doc.ufuncs",
+    "dtypes",
     "exceptions",
     "f2py",
     "fft",
@@ -160,7 +161,6 @@ PUBLIC_MODULES = ['numpy.' + s for s in [
     "random",
     "testing",
     "testing.overrides",
-    "types",
     "typing",
     "typing.mypy_plugin",
     "version",
